Generating random decision and sample size looks for decision1 == FALSE or default option

A helper function for ocPredprobDist() to generate numeric of decisions decisions and random looks all_sizes for decision1 == TRUE.

Usage

h_decision_two_predprobDist(
  nnE,
  nnF,
  nnr,
  truep,
  xS,
  nS,
  parE,
  parS,
  tT,
  tF,
  phiU,
  phiFu,
  deltaE,
  deltaF,
  weights,
  weightsS,
  relativeDelta
)

Arguments

nnE

(numeric):
sample size or sizes where study can be stopped for Efficacy decision. If 0 or NULL and length(nnE) = 1 then no Efficacy looks are performed.

nnF

(numeric):
sample size or sizes where study can be stopped for Efficacy decision. If 0 or NULL and length(nnF) = 1 then no Futility looks are performed.

nnr

(numeric):
union of nnE and nnF.

truep

(number):
assumed true response rate or true rate (scenario).

xS

(number):
number of successes in the S group at interim.

nS

(number):
number of patients in the E group at interim.

parE

(numeric):
alpha and beta parameters for the prior on the treatment population. Default set at alpha = 1, beta = 1, or uniform prior.

parS

(numeric):
parameters for beta distribution. If it is a matrix, it needs to have 2 columns, and each row corresponds to each component of a beta-mixture distribution for the S group.

tT

(number):
threshold of which assumed truep exceeds acceptable threshold of p0.

tF

(number):
threshold of which assumed truep does not exceed threshold of p1.

phiU

(number):
upper threshold on the predictive probability.

phiFu

(number):
upper threshold on the predictive probability.

deltaE

(number):
margin by which the response rate in the treatment group should be better than in the standard of care or control or S group in Efficacy looks only. Note that this can also be negative as well.

deltaF

(number):
margin by which the response rate in the treatment group should be better than in the standard of care or control or S group in Futility looks only. Note that this can also be negative as well.

weights

(numeric):
the mixture weights of the beta mixture prior.

weightsS

(numeric):
weights for the S group.

relativeDelta

(flag):
If TRUE, then a relativeDelta is used. Represents that a minimum response rate in magnitude of delta of the S non-responding patients. See postprobDist().

Value

A list with the following elements:

decision : decision flag with TRUE for Go, FALSE for Stop, NA for Gray zone.

all_sizes : resulting number of look size, anything below maximum look size is an indicated interim, Futility or Efficacy or both.