Generating random decision and sample size looks

A helper function for ocPostprob() to generate numeric of decisions decisions and random looks all_sizes.

Usage

h_get_decision(nnr, truep, p0, p1, parE = c(1, 1), nnE, nnF, tL, tU)

Arguments

nnr

(numeric):
union of nnEand nnF.

truep

(number):
assumed true response rate.

p0

(number):
lower Futility threshold of response rate.

p1

(number):
upper Efficacy threshold of response rate.

parE

(numeric):
alpha and beta parameters for the prior on the treatment population. Default set at alpha = 1, beta = 1, or uniform prior.

nnE

(numeric):
sample size or sizes where study can be stopped for Efficacy decision. If different for Futility decision, specify in nnF.

nnF

(numeric):
sample size or sizes where study can be stopped for Futility decision if different from Efficacy decision.

tL

(number):
posterior probability threshold for being below p0.

tU

(number):
posterior probability threshold for being above p1.

Value

A list with the following elements :

decision : decision flag with TRUE for Go, FALSE for Stop, NA for Gray zone.

all_sizes : resulting number of look size, anything below maximum look size is an indicated interim, Futility or Efficacy or both.