Generating random decision and sample size looks

A helper function for ocRctPostprobDist() to generate numeric of decisions decisions and random looks all_sizes.

Usage

h_get_decisionDist_rct(
  nnr,
  nnrE,
  nnrF,
  pE,
  pS,
  parE = c(1, 1),
  parS = c(1, 1),
  tL,
  tU,
  deltaE,
  deltaF,
  relativeDelta,
  randRatio = 1,
  Nmax,
  orig_nnr
)

Arguments

nnr

(numeric):
union of nnE and nnF.

parE

(numeric):
alpha and beta parameters for the prior on the treatment population. Default set at alpha = 1, beta = 1, or uniform prior.

randRatio

(numeric):
The randomisation ratio between treatment and control. Must be greater than 0 and maximum of 1.

Nmax

(number):
The max sample size or the sample size of final look.

orig_nnr

(numeric):
unique values of original, user-input values of nnE and nnF, defined by user-facing function and aid to plotOc()'s x axis input when wiggle = TRUE.

Value

A list with the following elements :

decision : decision flag with TRUE for Go, FALSE for Stop, NA for Gray zone.

all_sizes : resulting number of look size, anything below maximum look size is an indicated interim, Futility or Efficacy or both.

nActive : mean of look size for Active arm.

nControl : mean of look size for Control arm.