Generating random decision and sample size looks for decision1 == FALSE

A helper function for ocPredprob() to generate numeric of decisions decisions and random looks all_sizes for decision1 == FALSE.

Usage

h_get_decision_two_predprob(
  nnr,
  truep,
  p0,
  p1,
  parE = c(1, 1),
  nnE,
  nnF,
  tT,
  tF,
  phiFu,
  phiU
)

Arguments

nnr

(numeric):
union of nnE and nnF.

truep

(number):
assumed true response rate or true rate (scenario).

p0

(number):
lower Futility threshold of response rate.

p1

(number):
upper Futility threshold of response rate.

parE

(numeric):
alpha and beta parameters for the prior on the treatment population. Default set at alpha = 1, beta = 1, or uniform prior.

nnE

(numeric):
sample size or sizes where study can be stopped for Efficacy decision. If 0 or NULL and length(nnE) = 1 then no Efficacy looks are performed.

nnF

(numeric):
sample size or sizes where study can be stopped for Efficacy decision. If 0 or NULL and length(nnF) = 1 then no Futility looks are performed.

tT

(number):
threshold of which assumed truep exceeds acceptable threshold of p0.

tF

(number):
threshold of which assumed truep does not exceed threshold of p1.

phiFu

(number):
upper threshold on the predictive probability.

phiU

(number):
upper threshold on the predictive probability.

Value

A list with the following elements:

decision : decision flag with TRUE for Go, FALSE for Stop, NA for Gray zone.

all_sizes : resulting number of look size, anything below maximum look size is an indicated interim, Futility or Efficacy or both.