Generating looks from random distance

A helper function for ocPostprob() that applies the numeric element of dist to looks nn.

Usage

h_get_looks(dist, nnE, nnF)

Arguments

dist

(numeric or logical):
distance for random looks around the look locations in nn, where dist is generated from h_get_distance() in a numeric of at least one element. If NULL, only one location look will be set at nnE or nnF.

nnE

(numeric):
sample size or sizes where study can be stopped for Efficacy decision. If different for Futility decision, specify in nnF.

nnF

(numeric):
sample size or sizes where study can be stopped for Futility decision if different from Efficacy decision.

Value

Uses distance from h_get_distance() to add to looks, creating wiggled looks:

nnrE is the result for Efficacy looks with random distance added.

nnrF is the result for Futility looks with random distance added.