The predictive probability of success in two-arm studies.

The helper function to generate the predictive probability of success based on the difference in treatment group (E) and control or standard of care (S) group.

Usage

h_predprobdist(
  x,
  n,
  xS,
  nS,
  Nmax,
  NmaxControl,
  delta,
  relativeDelta,
  parE,
  parS,
  weights,
  weightsS,
  thetaT
)

Arguments

x

(number):
number of successes in the E group at interim.

n

(number):
number of patients in the E group at interim.

xS

(number):
number of successes in the S group at interim.

nS

(number):
number of patients in the E group at interim.

Nmax

(number):
maximum number of patients in the E group at final analysis.

NmaxControl

(number):
maximum number of patients in the S group at final analysis.

delta

(number):
difference between response rates to be met.

relativeDelta

(flag):
If TRUE, then a relativeDelta is used. Represents that a minimum response rate in magnitude of delta of the S non-responding patients. See postprobDist().

parE

(numeric):
parameters for beta distribution. If it is a matrix, it needs to have 2 columns, and each row corresponds to each component of a beta-mixture distribution for the E group.

parS

(numeric):
parameters for beta distribution. If it is a matrix, it needs to have 2 columns, and each row corresponds to each component of a beta-mixture distribution for the S group.

weights

(numeric):
the mixture weights of the beta mixture prior.

weightsS

(numeric):
weights for the S group.

thetaT

(number):
threshold on the probability to be used.

Value

A list is returned with names result for predictive probability and table of numeric values with counts of responses in the remaining patients, density for probabilities of these counts, posterior for corresponding probabilities to be above threshold, and successfor trial success indicators.