Compute the posterior probability with beta prior on SOC

Using the approach by Thall and Simon (Biometrics, 1994), evaluate the posterior probability of having Pr(P_E > P_S + delta | data) (but see below for relative delta margin). Both for the new treatment E as well as for the SOC S data might be available. However the default is that no data is available for the SOC, corresponding to the single arm trial situation. Note that a uniform prior is the useful default for the treatment proportion, while in the single arm trial an informative prior on the SOC proportion is useful.

Usage

postprobDist(
  x,
  n,
  xS = 0,
  nS = 0,
  delta = 0,
  relativeDelta = FALSE,
  parE = c(1, 1),
  weights,
  parS = c(1, 1),
  weightsS
)

Arguments

x

number of successes (in the treatment group). Note that x can be a vector.

n

number of patients (in the treatment group)

xS

number of successes in the SOC group (default: 0)

nS

number of patients in the SOC group (default: 0)

delta

margin by which the response rate in the treatment group should be better than in the SOC group (default: 0)

relativeDelta

should the delta be relative? (not default). If this is TRUE, then a relative delta is used. This means we want to have response at least in delta proportion of the SOC non-responding patients. Non-responding patients rate is 1 - P_S, and we want to have P_S + (1 - P_S)

delta response rate (at least) in the treatment. That is, we evaluate the posterior probability Pr(P_E > P_S + (1 - P_S) * delta | data).

parE

the beta parameters matrix, with K rows and 2 columns, corresponding to the beta parameters of the K components. default is a uniform prior.

weights

the mixture weights of the beta mixture prior. Default are uniform weights across mixture components.

parS

beta parameters for the SOC group (default: uniform)

weightsS

weights for the SOC group (default: uniform)

Value

the posterior probability

Details

Beta mixture prior can be specified for the treatment (parE and weights parameters) and control proportion (parS and weightsS parameters), see postprob for details. Note that being able to specify a beta mixture prior also on the control treatment is e.g. important for the futility decision making (see the oc2 code).

Examples

## example similar to Lee and Liu:
postprobDist(x = 16, n = 23, parE = c(0.6, 0.4), parS = c(0.6, 0.4))
#> [1] 0.5123873

## these two should give the same result:
postprobDist(
  x = 27, n = 34,
  xS = 0, nS = 0,
  delta = 0.15,
  parE = c(1, 1),
  parS = c(50007530, 49924090)
)
#> [1] 0.9576104

postprob(x = 27, n = 34, p = 0.65, parE = c(1, 1))
#> [1] 0.9580618
## ok, almost

## try out mixtures:
## play around with the beta parameters and weights to
## get a feeling.
## Note that very extreme beta parameters do no longer increase
## the return value, because then that mixture component is too
## unlikely a posteriori
postprobDist(
  x = 16, n = 23,
  parE =
    rbind(
      c(0.6, 0.4),
      c(10, 20)
    ),
  parS =
    rbind(
      c(0.6, 0.4),
      c(10, 10)
    ),
  weightsS = c(1, 3)
)
#> [1] 0.7681603