Predictive probability of trial success

Compute the predictive probability of trial success given current data. Success means that at the end of the trial the posterior probability is Pr(P_E > p) >= thetaT, where p is the fixed response rate threshold. Then the predictive probability for success is: pp = sum over i: Pr(Y=i|x,n)*I{Pr(P_E > p|x,Y=i)>=thetaT} where Y is the number of future responses in the treatment group and x is the current number of responses in the treatment group out of n. The prior for the response rate in the experimental arm is P_E ~ beta(a, b).

A table with the following contents will be included in the return output :

i: Y = i, number of future successes in Nmax-n subjects.

density: Pr(Y = i|x) using beta-(mixture)-binomial distribution.

posterior: Pr(P_E > p | x, Y=i) using beta posterior.

success: indicator I(b>thetaT).

Usage

predprob(x, n, Nmax, p, thetaT, parE = c(1, 1), weights)

Arguments

x

(number):
number of successes.

n

(number):
number of patients.

Nmax

(number):
maximum number of patients at the end of the trial.

p

(number):
threshold on the response rate.

thetaT

(number):
threshold on the probability to be above p.

parE

(numeric):
the beta parameters matrix, with K rows and 2 columns, corresponding to the beta parameters of the K components.

weights

(numeric):
the mixture weights of the beta mixture prior.

Value

A list is returned with names result for predictive probability and table of numeric values with counts of responses in the remaining patients, probabilities of these counts, corresponding probabilities to be above threshold, and trial success indicators.

References

Lee, J. J., & Liu, D. D. (2008). A predictive probability design for phase II cancer clinical trials. Clinical Trials, 5(2), 93-106. doi:10.1177/1740774508089279

Examples

# The original Lee and Liu (Table 1) example:
# Nmax = 40, x = 16, n = 23, beta(0.6,0.4) prior distribution,
# thetaT = 0.9. The control response rate is 60%:
predprob(
  x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.9,
  parE = c(0.6, 0.4)
)
#> Error in predprob(x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.9, parE = c(0.6,     0.4)): object 'posterior' not found

# Lowering/Increasing the probability threshold thetaT of course increases
# /decreases the predictive probability of success, respectively:
predprob(
  x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.8,
  parE = c(0.6, 0.4)
)
#> Error in predprob(x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.8, parE = c(0.6,     0.4)): object 'posterior' not found

predprob(
  x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.95,
  parE = c(0.6, 0.4)
)
#> Error in predprob(x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.95, parE = c(0.6,     0.4)): object 'posterior' not found

# Mixed beta prior
predprob(
  x = 20, n = 23, Nmax = 40, p = 0.6, thetaT = 0.9,
  parE = rbind(c(1, 1), c(25, 15)),
  weights = c(3, 1)
)
#> Error in predprob(x = 20, n = 23, Nmax = 40, p = 0.6, thetaT = 0.9, parE = rbind(c(1,     1), c(25, 15)), weights = c(3, 1)): object 'posterior' not found