Compute the predictive probability that the trial will be successful, with a prior distribution on the SOC

Compute the predictive probability of trial success given current data. Success means that at the end of the trial the posterior probability Pr(P_E > P_S + delta_0 | data) >= thetaT. Then the predictive probability for success is: pp = sum over i: Pr(Y=i|x,n)*IPr(P_E > P_S + delta|x,Y=i)>=thetaT, where Y is the number of future responses in the treatment group and x is the current number of responses in the treatment group (out of n). Prior is P_E ~ beta(a, b), default uniform which is a beta(1,1). However, also a beta mixture prior can be specified. Analogously for P_S either a classic beta prior or a beta mixture prior can be specified.

Usage

predprobDist(
  x,
  n,
  xS = 0,
  nS = 0,
  Nmax,
  NmaxControl = 0,
  delta = 0,
  relativeDelta = FALSE,
  parE = c(a = 1, b = 1),
  weights,
  parS = c(a = 1, b = 1),
  weightsS,
  thetaT
)

Arguments

x

number of successes (in the treatment group)

n

number of patients (in the treatment group)

xS

number of successes in the SOC group (default: 0)

nS

number of patients in the SOC group (default: 0)

Nmax

maximum number of patients at the end of the trial (in the treatment group)

NmaxControl

maximum number of patients at the end of the trial in the SOC group (default: 0)

delta

margin by which the response rate in the treatment group should be better than in the SOC group (default: 0)

relativeDelta

see postprobDist

parE

the beta parameters matrix, with K rows and 2 columns, corresponding to the beta parameters of the K components. default is a uniform prior.

weights

the mixture weights of the beta mixture prior. Default are uniform weights across mixture components.

parS

beta prior parameters in the SOC group (default: uniform)

weightsS

weights for the SOC group (default: uniform)

thetaT

threshold on the probability to be used

Value

The predictive probability, a numeric value. In addition, a list called tables is returned as attribute of the returned number.

Details

Also data on the SOC might be available. Then the predictive probability is more generally defined as pp = sum over i, j: Pr(Y=i|x,n)*Pr(Z=j|xS, nS)*IPr(P_E > P_S + delta|x,xS,Y=i,Z=j)>=thetaT where Z is the future number of responses in the SOC group, and xS is the current number of responses in the SOC group.

A table with the following contents will be included in the tables attribute of the return value (in the case that NmaxControl is zero): i: Y=i (number of future successes in Nmax-n subjects) py: Pr(Y=i|x) using beta-(mixture)-binomial distribution b: Pr(P_E > P_S + delta | x, Y=i) bgttheta: indicator I(b>thetaT)

If NmaxControl is not zero, i.e., when data on the control treatment is available in this trial, then a list with will be included with the following elements: pyz: matrix with the probabilities Pr(Y=i, Z=j | x, xS) b: matrix with Pr(P_E > P_S + delta | x, xS, Y=i, Z=j) bgttheta: matrix of indicators I(b>thetaT)

References

Lee, J. J., & Liu, D. D. (2008). A predictive probability design for phase II cancer clinical trials. Clinical Trials, 5(2), 93-106. doi:10.1177/1740774508089279

Examples

## The original Lee and Liu (Table 1) example:
## Nmax=40, x=16, n=23, beta(0.6,0.4) prior distribution,
## thetaT=0.9. The control response rate is 60%:
predprob(
  x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.9,
  parE = c(0.6, 0.4)
)
#> Error in predprob(x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.9, parE = c(0.6,     0.4)): object 'posterior' not found
## So the predictive probability is 56.6%. 12 responses
## in the remaining 17 patients are required for success.

## Now to the modified example, where the control response
## rate p has a beta(7, 11) distribution.
## For example if the 60% threshold stems from a trial
## with 10 patients, of which 6 were observed with response.
predprobDist(
  x = 16, n = 23, Nmax = 40, delta = 0, thetaT = 0.9,
  parE = c(0.6, 0.4), parS = c(7, 11)
)
#> [1] 0.9778204
#> attr(,"tables")
#>        i     py      b bgttheta
#>  [1,]  0 0.0000 0.5531        0
#>  [2,]  1 0.0000 0.6215        0
#>  [3,]  2 0.0001 0.6857        0
#>  [4,]  3 0.0006 0.7444        0
#>  [5,]  4 0.0021 0.7967        0
#>  [6,]  5 0.0058 0.8420        0
#>  [7,]  6 0.0135 0.8802        0
#>  [8,]  7 0.0276 0.9115        1
#>  [9,]  8 0.0497 0.9365        1
#> [10,]  9 0.0794 0.9558        1
#> [11,] 10 0.1129 0.9702        1
#> [12,] 11 0.1426 0.9806        1
#> [13,] 12 0.1587 0.9879        1
#> [14,] 13 0.1532 0.9928        1
#> [15,] 14 0.1246 0.9959        1
#> [16,] 15 0.0811 0.9978        1
#> [17,] 16 0.0381 0.9989        1
#> [18,] 17 0.0099 0.9995        1
## Now only 11 responses are needed in the remaining 17 patients
## in order to have a successful trial.
## Also, the predictive probability for a Go decision
## is now 70.8% instead of merely 56.6%.

## If the response threshold of 60% stems from an SOC estimate (e.g. 50% = 5/10)
## and a margin of e.g. 10%, then the margin can be specified with
## the delta argument:
predprobDist(
  x = 16, n = 23, Nmax = 40, delta = 0.1, thetaT = 0.9,
  parE = c(1, 1), parS = c(6, 11)
)
#> [1] 0.8004621
#> attr(,"tables")
#>        i     py      b bgttheta
#>  [1,]  0 0.0000 0.3716        0
#>  [2,]  1 0.0000 0.4401        0
#>  [3,]  2 0.0002 0.5092        0
#>  [4,]  3 0.0007 0.5773        0
#>  [5,]  4 0.0025 0.6424        0
#>  [6,]  5 0.0068 0.7032        0
#>  [7,]  6 0.0157 0.7585        0
#>  [8,]  7 0.0315 0.8076        0
#>  [9,]  8 0.0555 0.8500        0
#> [10,]  9 0.0867 0.8859        0
#> [11,] 10 0.1203 0.9153        1
#> [12,] 11 0.1476 0.9388        1
#> [13,] 12 0.1589 0.9571        1
#> [14,] 13 0.1477 0.9709        1
#> [15,] 14 0.1151 0.9809        1
#> [16,] 15 0.0714 0.9880        1
#> [17,] 16 0.0317 0.9928        1
#> [18,] 17 0.0077 0.9959        1
## This again changes the result: Only 10 future responses are required,
## the predictive probability of success is now 80%.

## Next we will use a beta mixture prior, with a weight of 1/3 for an
## informative beta(50, 10) prior, on the experimental response rate:
predprobDist(
  x = 16, n = 23, Nmax = 40, delta = 0.1, thetaT = 0.9,
  parE = rbind(c(1, 1), c(50, 10)),
  weights = c(2, 1), parS = c(6, 11)
)
#> [1] 0.8719377
#> attr(,"tables")
#>        i     py      b bgttheta
#>  [1,]  0 0.0000 0.3716        0
#>  [2,]  1 0.0000 0.4402        0
#>  [3,]  2 0.0001 0.5095        0
#>  [4,]  3 0.0004 0.5777        0
#>  [5,]  4 0.0015 0.6433        0
#>  [6,]  5 0.0041 0.7050        0
#>  [7,]  6 0.0095 0.7617        0
#>  [8,]  7 0.0194 0.8128        0
#>  [9,]  8 0.0352 0.8579        0
#> [10,]  9 0.0578 0.8966        0
#> [11,] 10 0.0868 0.9284        1
#> [12,] 11 0.1194 0.9532        1
#> [13,] 12 0.1495 0.9710        1
#> [14,] 13 0.1661 0.9828        1
#> [15,] 14 0.1566 0.9901        1
#> [16,] 15 0.1166 0.9943        1
#> [17,] 16 0.0605 0.9967        1
#> [18,] 17 0.0164 0.9980        1
## Now still 10 future responses are required for success, but the predictive
## success probability is higher (87.2%).

## We can also have additional controls in our trial. For example
## assume that in the interim analysis we had 10 control patients observed
## with 5 responses out of 20 in total at the end.
## We also had historical control data (say 60 patients of
## which 20 responded), but in order to be robust against changes in the response
# rate over time, we will again use a beta mixture, putting only 1/3 of weight
## on the historical controls:
predprobDist(
  x = 16, n = 23, xS = 5, nS = 10,
  Nmax = 40, NmaxControl = 20,
  delta = 0.1, thetaT = 0.9,
  parE = rbind(c(1, 1), c(50, 10)),
  weights = c(2, 1),
  parS = rbind(c(1, 1), c(20, 40)),
  weightsS = c(2, 1)
)
#> Error in uniroot(f, lower = 0, upper = 1): f() values at end points not of opposite sign
## Now we only have 59.9% predictive probability of success. The bgttheta matrix
## lists the cases where we would have success, for all possible combinations
## of future responses in the control and experimental arms. We see that the
## success threshold for experimental responses depends on the number of
## control responses.