Beta-Mixture Quantile Function

Calculates the quantile of the Beta-Mixture distribution for a given probability.

Usage

qbetaMix(p, par, weights, lower.tail = TRUE)

Arguments

p

(numeric):
the required probability.

par

(number):
the beta parameters matrix, with K rows and 2 columns, corresponding to the beta parameters of the K components.

weights

(matrix):
the mixture weights of the beta mixture prior.

lower.tail

(flag):
whether CDF at x taken at lower or upper tail

Value

The abscissa.

Examples

# Only 1 mixture component, i.e., weights = 1
qbetaMix(
  p = 0.60,
  par = rbind(c(0.2, 0.4)),
  weights = 1
)
#> [1] 0.3112068

# With 2 mixture components
qbetaMix(
  p = 0.6, par = rbind(c(0.2, 0.4), c(1, 1)),
  weights = c(0.6, 0.4)
)
#> [1] 0.488759

# Can also specify q as a vector
qbetaMix(
  p = seq(0, 1, .01),
  par = rbind(c(0.2, 0.4), c(1, 1)),
  weights = c(0.6, 0.4)
)
#>   [1] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 1.062599e-04
#>   [6] 1.195146e-04 1.155159e-04 8.443751e-05 1.848755e-04 3.796376e-04
#>  [11] 6.207540e-04 9.948499e-04 1.512693e-03 2.212278e-03 3.161665e-03
#>  [16] 4.396418e-03 5.976652e-03 7.922461e-03 1.032535e-02 1.315398e-02
#>  [21] 1.651358e-02 2.040127e-02 2.487029e-02 2.987466e-02 3.548362e-02
#>  [26] 4.166591e-02 4.842761e-02 5.573528e-02 6.364744e-02 7.204244e-02
#>  [31] 8.098204e-02 9.040336e-02 1.002934e-01 1.106289e-01 1.213959e-01
#>  [36] 1.325708e-01 1.440653e-01 1.559580e-01 1.681759e-01 1.807060e-01
#>  [41] 1.935289e-01 2.066262e-01 2.199802e-01 2.335744e-01 2.473929e-01
#>  [46] 2.614459e-01 2.756684e-01 2.900724e-01 3.046450e-01 3.193743e-01
#>  [51] 3.342283e-01 3.492379e-01 3.643711e-01 3.796162e-01 3.949599e-01
#>  [56] 4.103981e-01 4.259220e-01 4.415224e-01 4.571909e-01 4.729799e-01
#>  [61] 4.887590e-01 5.045812e-01 5.204381e-01 5.363217e-01 5.522238e-01
#>  [66] 5.681361e-01 5.840505e-01 5.999584e-01 6.158512e-01 6.317201e-01
#>  [71] 6.475557e-01 6.633485e-01 6.790885e-01 6.947652e-01 7.103673e-01
#>  [76] 7.258831e-01 7.412998e-01 7.565763e-01 7.717931e-01 7.868140e-01
#>  [81] 8.016817e-01 8.163618e-01 8.308578e-01 8.451166e-01 8.591210e-01
#>  [86] 8.728399e-01 8.862383e-01 8.992760e-01 9.119066e-01 9.240762e-01
#>  [91] 9.357219e-01 9.467694e-01 9.571314e-01 9.667238e-01 9.753373e-01
#>  [96] 9.829571e-01 9.893392e-01 9.943267e-01 9.977525e-01 9.995464e-01
#> [101] 1.000000e+00