Operating Characteristics for Posterior Probability method

Calculate operating characteristics for posterior probability method.

It is assumed that the true response rate is truep. The trial is stopped for Efficacy if the posterior probability to be above p1 is larger than tU, and stopped for Futility if the posterior probability to be below p0 is larger than tL:

Stop criteria for Efficacy :

Pr(truep > p1) > tU

Stop criteria for Futility :

Pr(truep < p0) > tL

Resulting operating characteristics include the following:

ExpectedN: expected number of patients in the trials

PrStopEarly: probability to stop the trial early (before reaching the maximum sample size)

PrEarlyEff: probability of Early Go decision

PrEarlyFut: probability of for Early Stop decision

PrEfficacy: probability of Go decision

PrFutility: probability of Stop decision

PrGrayZone: probability between Go and Stop ,"Evaluate" or Gray decision zone

Usage

ocPostprob(
  nnE,
  truep,
  p0,
  p1,
  tL,
  tU,
  parE = c(1, 1),
  sim = 50000,
  wiggle = FALSE,
  nnF = nnE
)

Arguments

nnE

(numeric):
sample size or sizes where study can be stopped for Efficacy decision. If different for Futility decision, specify in nnF.

truep

(number):
assumed true response rate.

p0

(number):
lower Futility threshold of response rate.

p1

(number):
upper Efficacy threshold of response rate.

tL

(number):
posterior probability threshold for being below p0.

tU

(number):
posterior probability threshold for being above p1.

parE

(numeric):
alpha and beta parameters for the prior on the treatment population. Default set at alpha = 1, beta = 1, or uniform prior.

sim

(number):
number of simulations.

wiggle

(flag):
generate random look locations (not default). if TRUE, optional to specify dist (see @details).

nnF

(numeric):
sample size or sizes where study can be stopped for Futility decision if different from Efficacy decision.

Value

A list with the following elements:

oc: matrix with operating characteristics (see @details section)

nn: vector of look locations that was supplied

nnE: vector of Efficacy look locations

nnF: vector of Futility look locations

params: user input arguments.

Examples

# For three looks of 10, 20 and 30 we have the following assumptions :
# True response rate or truep of the treatment group = 40%
# The following are the Stop rules respectively :
# Look for Efficacy: Pr(truep > 30% )> 80%
# Look for Futility: Pr(truep < 20% )> 60%
# Prior of treatment arm parE= Beta(1,1).
res <- ocPostprob(
  nnE = c(10, 20, 30), truep = 0.40, p0 = 0.20, p1 = 0.30, tL = 0.60, tU = 0.80, parE = c(1, 1),
  sim = 50000, wiggle = TRUE, nnF = c(10, 20, 30)
)

res$oc
#>   ExpectedN PrStopEarly PrEarlyEff PrEarlyFut PrEfficacy PrFutility PrGrayZone
#> 1   18.1194     0.69226    0.65074    0.04152     0.7724    0.04356    0.18404

# We specify the distance in this example.
res <- ocPostprob(
  nnE = c(10, 20, 30), truep = 0.40, p0 = 0.20, p1 = 0.30, tL = 0.60, tU = 0.80, parE = c(1, 1),
  sim = 50000, wiggle = TRUE, nnF = c(10, 20, 30)
)

res$oc
#>   ExpectedN PrStopEarly PrEarlyEff PrEarlyFut PrEfficacy PrFutility PrGrayZone
#> 1  18.14264     0.69132    0.64888    0.04244    0.77192    0.04464    0.18344

# Here, nnE = nnF, and no wiggle room is allowed. Random distance also not supplied.
res <- ocPostprob(
  nnE = c(10, 20, 30), truep = 0.40, p0 = 0.20, p1 = 0.30, tL = 0.60, tU = 0.80, parE = c(1, 1),
  sim = 50000, wiggle = FALSE, nnF = c(10, 20, 30)
)

res$oc
#>   ExpectedN PrStopEarly PrEarlyEff PrEarlyFut PrEfficacy PrFutility PrGrayZone
#> 1   19.0732     0.67542    0.62196    0.05346    0.75414    0.05524    0.19062

# Here, we only have one Futility and Efficacy look or stop.
res <- ocPostprob(
  nnE = c(10), truep = 0.40, p0 = 0.20, p1 = 0.30, tL = 0.60, tU = 0.80, parE = c(1, 1),
  sim = 50000, wiggle = FALSE, nnF = 10
)

res$oc
#>   ExpectedN PrStopEarly PrEarlyEff PrEarlyFut PrEfficacy PrFutility PrGrayZone
#> 1        10           0          0          0    0.36634    0.04402    0.58964

# Here, we only have one Futility but many Efficacy looks or stop.
res <- ocPostprob(
  nnE = c(10, 20, 30), truep = 0.40, p0 = 0.20, p1 = 0.30, tL = 0.60, tU = 0.80, parE = c(1, 1),
  sim = 50000, wiggle = FALSE, nnF = 10
)

res$oc
#>   ExpectedN PrStopEarly PrEarlyEff PrEarlyFut PrEfficacy PrFutility PrGrayZone
#> 1   19.1642     0.66888    0.62092    0.04796    0.75338    0.04796    0.19866