Calculate operating characteristics for posterior probability method with beta prior on SOC

The trial is stopped for efficacy if the posterior probability to be at least deltaE better than the control is larger than tU, and stopped for futility if the posterior probability to be at least deltaF worse than the control is larger than tL. Otherwise the trial is continued, and at the maximum sample size it may happen that no decision is made ("gray zone").

Usage

ocPostprobDist(
  nn,
  p,
  deltaE,
  deltaF,
  relativeDelta = FALSE,
  tL,
  tU,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  ns = 10000,
  nr = FALSE,
  d = NULL,
  nnF = nn
)

Arguments

nn

vector of look locations for efficacy (if futility looks should be different, please specify also nnF)

p

true rate (scenario)

deltaE

delta for efficacy: P(pE > pS + deltaE) should be large to stop for efficacy

deltaF

delta for futility: P(pE < pS - deltaF) should be large to stop for futility

relativeDelta

see postprobDist

tL

probability threshold for being below control - deltaF

tU

probability threshold for being above control + deltaE

parE

beta parameters for the prior on the treatment proportion

parS

beta parameters for the prior on the control proportion

ns

number of simulations

nr

generate random look locations? (not default)

d

distance for random looks around the look locations in nn

nnF

vector of look locations for futility (default: same as efficacy)

Value

A list with the following elements: oc: matrix with operating characteristics (see Details section) Decision: vector of the decisions made in the simulated trials (TRUE for success, FALSE for failure, NA for no decision) SampleSize: vector of the sample sizes in the simulated trials nn: vector of look locations nnE: vector of efficacy look locations nnF: vector of futility look locations todo: would we like to return nnr instead, the actual look locations? params: input parameters for this function

Details

Returned operating characteristics in a matrix include: ExpectedN: expected number of patients in the trials PrStopEarly: probability to stop the trial early (before reaching the maximum sample size) PrEarlyEff: probability to decide for efficacy early PrEarlyFut: probability to decide for futility early PrEfficacy: probability to decide for efficacy PrFutility: probability to decide for futility PrGrayZone: probability of no decision at the end ("gray zone")

Examples

# operating characteristics for posterior probability method with beta prior on SOC

# design details
# multiple looks @ 10, 20, 30 patietns
# True response rate of the treatment group=0.4

# stop for efficacy (deltaE): P(pE > pS + deltaE) >tU
# stop for futility (deltaF): P(pE < pS - deltaF) >tL
# where pE is the response rate of treatment group, pS is the response rate of
# control group. Both of them are random variables.

res1 <- ocPostprobDist(
  nn = c(10, 20, 30),
  p = 0.4,
  deltaE = 0.1,
  deltaF = -0.1,
  tL = 0.6,
  tU = 0.6,
  parE = c(1, 1),
  parS = c(5, 25),
  ns = 100
)
#> Error in (function (x, n, xS = 0, nS = 0, delta, relativeDelta = FALSE,     parE = c(1, 1), weights, parS = c(1, 1), weightsS) {    if (is.vector(parE)) {        assert_true(identical(length(parE), 2L))        parE <- t(parE)    }    if (is.vector(parS)) {        assert_true(identical(length(parS), 2L))        parS <- t(parS)    }    if (missing(weights)) {        weights <- rep(1, nrow(parE))    }    if (missing(weightsS)) {        weightsS <- rep(1, nrow(parS))    }    assert_numeric(x, lower = 0, upper = n, finite = TRUE)    assert_number(n, lower = x, finite = TRUE)    assert_number(xS, lower = 0, upper = n, finite = TRUE)    assert_numeric(nS, lower = 0, finite = TRUE)    assert_number(delta, lower = 0, finite = TRUE)    assert_flag(relativeDelta)    assert_numeric(weights, lower = 0, finite = TRUE)    assert_numeric(weightsS, lower = 0, finite = TRUE)    assert_numeric(parE, lower = 0, finite = TRUE)    assert_numeric(parS, lower = 0, finite = TRUE)    activeBetamixPost <- h_getBetamixPost(x = x, n = n, par = parE,         weights = weights)    controlBetamixPost <- h_getBetamixPost(x = xS, n = nS, par = parS,         weights = weightsS)    assert_names(names(activeBetamixPost), identical.to = c("par",         "weights"))    assert_names(names(controlBetamixPost), identical.to = c("par",         "weights"))    if (relativeDelta) {        epsilon <- .Machine$double.xmin        integrand <- h_integrand_relDelta    }    else {        epsilon <- .Machine$double.xmin        integrand <- h_integrand    }    bounds <- h_get_bounds(controlBetamixPost = controlBetamixPost)    intRes <- integrate(f = integrand, lower = max(bounds[1],         ifelse(relativeDelta, 0, 0 - delta)), upper = min(ifelse(relativeDelta,         1, 1 - delta), bounds[2]), delta = delta, activeBetamixPost = activeBetamixPost,         controlBetamixPost = controlBetamixPost)    if (intRes$message == "OK") {        intRes$value    }    else {        stop(intRes$message)    }})(x = dots[[1L]][[1L]], n = 0, xS = 5L, nS = 10, delta = -0.1,     relativeDelta = FALSE, parE = c(5, 25), parS = c(1, 1)): Assertion on 'xS' failed: Element 1 is not <= 0.

res1$oc
#> Error in eval(expr, envir, enclos): object 'res1' not found

# generate random look locations around 10,20,30 patients
# this call will generate d (distance for random looks around the look locations)
# based on "floor(min(nn - c(0,nn[-length(nn)]))/2)" as d is missing:
res2 <- ocPostprobDist(
  nn = c(10, 20, 30),
  p = 0.4,
  deltaE = 0.1,
  deltaF = -0.1,
  tL = 0.6,
  tU = 0.6,
  parE = c(1, 1),
  parS = c(5, 25),
  ns = 100, nr = TRUE
)
#> Error in (function (x, n, xS = 0, nS = 0, delta, relativeDelta = FALSE,     parE = c(1, 1), weights, parS = c(1, 1), weightsS) {    if (is.vector(parE)) {        assert_true(identical(length(parE), 2L))        parE <- t(parE)    }    if (is.vector(parS)) {        assert_true(identical(length(parS), 2L))        parS <- t(parS)    }    if (missing(weights)) {        weights <- rep(1, nrow(parE))    }    if (missing(weightsS)) {        weightsS <- rep(1, nrow(parS))    }    assert_numeric(x, lower = 0, upper = n, finite = TRUE)    assert_number(n, lower = x, finite = TRUE)    assert_number(xS, lower = 0, upper = n, finite = TRUE)    assert_numeric(nS, lower = 0, finite = TRUE)    assert_number(delta, lower = 0, finite = TRUE)    assert_flag(relativeDelta)    assert_numeric(weights, lower = 0, finite = TRUE)    assert_numeric(weightsS, lower = 0, finite = TRUE)    assert_numeric(parE, lower = 0, finite = TRUE)    assert_numeric(parS, lower = 0, finite = TRUE)    activeBetamixPost <- h_getBetamixPost(x = x, n = n, par = parE,         weights = weights)    controlBetamixPost <- h_getBetamixPost(x = xS, n = nS, par = parS,         weights = weightsS)    assert_names(names(activeBetamixPost), identical.to = c("par",         "weights"))    assert_names(names(controlBetamixPost), identical.to = c("par",         "weights"))    if (relativeDelta) {        epsilon <- .Machine$double.xmin        integrand <- h_integrand_relDelta    }    else {        epsilon <- .Machine$double.xmin        integrand <- h_integrand    }    bounds <- h_get_bounds(controlBetamixPost = controlBetamixPost)    intRes <- integrate(f = integrand, lower = max(bounds[1],         ifelse(relativeDelta, 0, 0 - delta)), upper = min(ifelse(relativeDelta,         1, 1 - delta), bounds[2]), delta = delta, activeBetamixPost = activeBetamixPost,         controlBetamixPost = controlBetamixPost)    if (intRes$message == "OK") {        intRes$value    }    else {        stop(intRes$message)    }})(x = dots[[1L]][[1L]], n = 0, xS = 7L, nS = 10, delta = -0.1,     relativeDelta = FALSE, parE = c(5, 25), parS = c(1, 1)): Assertion on 'xS' failed: Element 1 is not <= 0.

res2$oc
#> Error in eval(expr, envir, enclos): object 'res2' not found

# specify the distance for random looks around the look locations in nn (d=5 for illustration)
res3 <- ocPostprobDist(
  nn = c(10, 20, 30),
  p = 0.4,
  deltaE = 0.1,
  deltaF = -0.1,
  tL = 0.6,
  tU = 0.6,
  parE = c(1, 1),
  parS = c(5, 25),
  ns = 100,
  nr = TRUE,
  d = 5
)
#> Error in (function (x, n, xS = 0, nS = 0, delta, relativeDelta = FALSE,     parE = c(1, 1), weights, parS = c(1, 1), weightsS) {    if (is.vector(parE)) {        assert_true(identical(length(parE), 2L))        parE <- t(parE)    }    if (is.vector(parS)) {        assert_true(identical(length(parS), 2L))        parS <- t(parS)    }    if (missing(weights)) {        weights <- rep(1, nrow(parE))    }    if (missing(weightsS)) {        weightsS <- rep(1, nrow(parS))    }    assert_numeric(x, lower = 0, upper = n, finite = TRUE)    assert_number(n, lower = x, finite = TRUE)    assert_number(xS, lower = 0, upper = n, finite = TRUE)    assert_numeric(nS, lower = 0, finite = TRUE)    assert_number(delta, lower = 0, finite = TRUE)    assert_flag(relativeDelta)    assert_numeric(weights, lower = 0, finite = TRUE)    assert_numeric(weightsS, lower = 0, finite = TRUE)    assert_numeric(parE, lower = 0, finite = TRUE)    assert_numeric(parS, lower = 0, finite = TRUE)    activeBetamixPost <- h_getBetamixPost(x = x, n = n, par = parE,         weights = weights)    controlBetamixPost <- h_getBetamixPost(x = xS, n = nS, par = parS,         weights = weightsS)    assert_names(names(activeBetamixPost), identical.to = c("par",         "weights"))    assert_names(names(controlBetamixPost), identical.to = c("par",         "weights"))    if (relativeDelta) {        epsilon <- .Machine$double.xmin        integrand <- h_integrand_relDelta    }    else {        epsilon <- .Machine$double.xmin        integrand <- h_integrand    }    bounds <- h_get_bounds(controlBetamixPost = controlBetamixPost)    intRes <- integrate(f = integrand, lower = max(bounds[1],         ifelse(relativeDelta, 0, 0 - delta)), upper = min(ifelse(relativeDelta,         1, 1 - delta), bounds[2]), delta = delta, activeBetamixPost = activeBetamixPost,         controlBetamixPost = controlBetamixPost)    if (intRes$message == "OK") {        intRes$value    }    else {        stop(intRes$message)    }})(x = dots[[1L]][[1L]], n = 0, xS = 5L, nS = 9, delta = -0.1,     relativeDelta = FALSE, parE = c(5, 25), parS = c(1, 1)): Assertion on 'xS' failed: Element 1 is not <= 0.

res3$oc
#> Error in eval(expr, envir, enclos): object 'res3' not found

# finally, we can also have separate specification of efficacy and futility analyses. E.g.
# here have futility decisions only at the final analysis:
res4 <- ocPostprobDist(
  nn = c(10, 20, 30), p = 0.4, deltaE = 0.1, deltaF = -0.1, tL = 0.6, tU = 0.6,
  parE = c(1, 1), parS = c(5, 25), ns = 100, nr = TRUE, d = 5, nnF = c(30)
)
#> Error in (function (x, n, xS = 0, nS = 0, delta, relativeDelta = FALSE,     parE = c(1, 1), weights, parS = c(1, 1), weightsS) {    if (is.vector(parE)) {        assert_true(identical(length(parE), 2L))        parE <- t(parE)    }    if (is.vector(parS)) {        assert_true(identical(length(parS), 2L))        parS <- t(parS)    }    if (missing(weights)) {        weights <- rep(1, nrow(parE))    }    if (missing(weightsS)) {        weightsS <- rep(1, nrow(parS))    }    assert_numeric(x, lower = 0, upper = n, finite = TRUE)    assert_number(n, lower = x, finite = TRUE)    assert_number(xS, lower = 0, upper = n, finite = TRUE)    assert_numeric(nS, lower = 0, finite = TRUE)    assert_number(delta, lower = 0, finite = TRUE)    assert_flag(relativeDelta)    assert_numeric(weights, lower = 0, finite = TRUE)    assert_numeric(weightsS, lower = 0, finite = TRUE)    assert_numeric(parE, lower = 0, finite = TRUE)    assert_numeric(parS, lower = 0, finite = TRUE)    activeBetamixPost <- h_getBetamixPost(x = x, n = n, par = parE,         weights = weights)    controlBetamixPost <- h_getBetamixPost(x = xS, n = nS, par = parS,         weights = weightsS)    assert_names(names(activeBetamixPost), identical.to = c("par",         "weights"))    assert_names(names(controlBetamixPost), identical.to = c("par",         "weights"))    if (relativeDelta) {        epsilon <- .Machine$double.xmin        integrand <- h_integrand_relDelta    }    else {        epsilon <- .Machine$double.xmin        integrand <- h_integrand    }    bounds <- h_get_bounds(controlBetamixPost = controlBetamixPost)    intRes <- integrate(f = integrand, lower = max(bounds[1],         ifelse(relativeDelta, 0, 0 - delta)), upper = min(ifelse(relativeDelta,         1, 1 - delta), bounds[2]), delta = delta, activeBetamixPost = activeBetamixPost,         controlBetamixPost = controlBetamixPost)    if (intRes$message == "OK") {        intRes$value    }    else {        stop(intRes$message)    }})(x = dots[[1L]][[1L]], n = 0, xS = 9L, nS = 30, delta = -0.1,     relativeDelta = FALSE, parE = c(5, 25), parS = c(1, 1)): Assertion on 'xS' failed: Element 1 is not <= 0.
res4$oc
#> Error in eval(expr, envir, enclos): object 'res4' not found
# compared to res3, we see that there is no early futility stopping anymore,
# and the overall probability for stopping for futility (which is the type II error here)
# is much lower.