Calculate operating characteristics for RCT against SOC, using the posterior probability method with beta priors

The randomization works as follows. According to the randomization ratio and the maximum sample size, patients are allocated to the treatment and SOC arms. The number of patients in the active treatment arm is rounded to the next higher integer. That is, the sequence of patients is determined from the start, such that the number of patients in both arms is constant across trial simulations.

Usage

ocRctPostprobDist(
  nn,
  pE,
  pS,
  deltaE,
  deltaF,
  relativeDelta = FALSE,
  tL = 0.8,
  tU = 0.8,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  randRatio = 1,
  ns = 10000,
  nr = FALSE,
  d = NULL,
  nnF = nn
)

Arguments

nn

vector of look locations for efficacy. Note that the maximum sample size is derived as the maximum value (latest look) in this vector. (if futility looks should be different, please specify also nnF)

pE

true rate in the treatment population

pS

true rate in the standard of care population

deltaE

delta for efficacy: P(pE > pS + deltaE) should be large to stop for efficacy. Must be non-negative value between 0 and 1.

deltaF

delta for futility: P(pE < pS - deltaF) should be large to stop for futility. Must be numeric value between -1 and 1. Can also be NULL. (see details)

relativeDelta

see postprobDist

tL

probability threshold for being below control - deltaF (default: 0.8)

tU

probability threshold for being above control + deltaE (default: 0.8)

parE

beta parameters for the prior on the control proportion

parS

beta parameters for the prior on the treatment proportion

randRatio

the randomization ratio (active vs. control) to be used. default: 1.

ns

number of simulations

nr

generate random look locations? (not default)

d

distance for random looks around the look locations in nn

nnF

vector of look locations for futility (default: same as efficacy)

Value

A list with the following elements: oc: matrix with operating characteristics (see Details section) Decision: vector of the decisions made in the simulated trials (TRUE for success, FALSE for failure, NA for no decision) SampleSize: vector of the sample sizes in the simulated trials SampleSizeActive: vector of the patients with treatment in the simulated trials SampleSizeControl: vector of the patients with control in the simulated trials nn: vector of look locations nnE: vector of efficacy look locations nnF: vector of futility look locations todo: would we like to return nnr instead, the actual look locations? params: input parameters for this function

Details

The randomized controlled trial (RCT) is stopped for efficacy if the posterior probability to be at least deltaE better than the control is larger than tU, and stopped for futility if the posterior probability to be at least deltaF worse than the control is larger than tL. Otherwise the trial is continued, and at the maximum sample size it may happen that no decision is made ("gray zone").

A variation can be requested when deltaF is set to NULL. Then the futility decision is made when the probability to be deltaE better than the control is lower than 1 - tL.

Returned operating characteristics in a matrix include: ExpectedN: expected total number of patients in the trials ExpectedNactive: expected number of patients with treatment ExpectedNcontrol: expected number of patients with SOC PrStopEarly: probability to stop the trial early (before reaching the maximum sample size) PrEfficacy: probability to decide for efficacy PrFutility: probability to decide for futility PrGrayZone: probability of no decision at the end ("gray zone")

Examples

# note, deltaF has to be a value between 0 and 1 including the bounds, or NULL.

res <- ocRctPostprobDist(
  nn = c(10, 20, 30), pE = 0.4, pS = 0.3, deltaE = 0.15, deltaF = 0.05,
  relativeDelta = FALSE,
  tL = 0.8, tU = 0.8,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  randRatio = 1,
  ns = 100, nr = FALSE, d = NULL
)
#> Error in (function (x, n, xS = 0, nS = 0, delta, relativeDelta = FALSE,     parE = c(1, 1), weights, parS = c(1, 1), weightsS) {    if (is.vector(parE)) {        assert_true(identical(length(parE), 2L))        parE <- t(parE)    }    if (is.vector(parS)) {        assert_true(identical(length(parS), 2L))        parS <- t(parS)    }    if (missing(weights)) {        weights <- rep(1, nrow(parE))    }    if (missing(weightsS)) {        weightsS <- rep(1, nrow(parS))    }    assert_numeric(x, lower = 0, upper = n, finite = TRUE)    assert_number(n, lower = x, finite = TRUE)    assert_number(xS, lower = 0, upper = n, finite = TRUE)    assert_numeric(nS, lower = 0, finite = TRUE)    assert_number(delta, lower = 0, finite = TRUE)    assert_flag(relativeDelta)    assert_numeric(weights, lower = 0, finite = TRUE)    assert_numeric(weightsS, lower = 0, finite = TRUE)    assert_numeric(parE, lower = 0, finite = TRUE)    assert_numeric(parS, lower = 0, finite = TRUE)    activeBetamixPost <- h_getBetamixPost(x = x, n = n, par = parE,         weights = weights)    controlBetamixPost <- h_getBetamixPost(x = xS, n = nS, par = parS,         weights = weightsS)    assert_names(names(activeBetamixPost), identical.to = c("par",         "weights"))    assert_names(names(controlBetamixPost), identical.to = c("par",         "weights"))    if (relativeDelta) {        epsilon <- .Machine$double.xmin        integrand <- h_integrand_relDelta    }    else {        epsilon <- .Machine$double.xmin        integrand <- h_integrand    }    bounds <- h_get_bounds(controlBetamixPost = controlBetamixPost)    intRes <- integrate(f = integrand, lower = max(bounds[1],         ifelse(relativeDelta, 0, 0 - delta)), upper = min(ifelse(relativeDelta,         1, 1 - delta), bounds[2]), delta = delta, activeBetamixPost = activeBetamixPost,         controlBetamixPost = controlBetamixPost)    if (intRes$message == "OK") {        intRes$value    }    else {        stop(intRes$message)    }})(x = dots[[1L]][[1L]], n = 3L, xS = 4L, nS = 7L, delta = 0.15,     relativeDelta = FALSE, parE = c(a = 1, b = 1), parS = c(a = 1,     b = 1)): Assertion on 'xS' failed: Element 1 is not <= 3.

res$oc
#> Error in eval(expr, envir, enclos): object 'res' not found

# separate specification of efficacy and futility analyses can be performed. E.g.
# here have futility decisions only at the final analysis:

res2 <- ocRctPostprobDist(
  nn = c(10, 20, 30), pE = 0.4, pS = 0.3, deltaE = 0.15, deltaF = 0.05,
  relativeDelta = FALSE,
  tL = 0.8, tU = 0.8,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  randRatio = 1,
  ns = 100, nr = FALSE, d = NULL, nnF = c(30)
)
#> Error in (function (x, n, xS = 0, nS = 0, delta, relativeDelta = FALSE,     parE = c(1, 1), weights, parS = c(1, 1), weightsS) {    if (is.vector(parE)) {        assert_true(identical(length(parE), 2L))        parE <- t(parE)    }    if (is.vector(parS)) {        assert_true(identical(length(parS), 2L))        parS <- t(parS)    }    if (missing(weights)) {        weights <- rep(1, nrow(parE))    }    if (missing(weightsS)) {        weightsS <- rep(1, nrow(parS))    }    assert_numeric(x, lower = 0, upper = n, finite = TRUE)    assert_number(n, lower = x, finite = TRUE)    assert_number(xS, lower = 0, upper = n, finite = TRUE)    assert_numeric(nS, lower = 0, finite = TRUE)    assert_number(delta, lower = 0, finite = TRUE)    assert_flag(relativeDelta)    assert_numeric(weights, lower = 0, finite = TRUE)    assert_numeric(weightsS, lower = 0, finite = TRUE)    assert_numeric(parE, lower = 0, finite = TRUE)    assert_numeric(parS, lower = 0, finite = TRUE)    activeBetamixPost <- h_getBetamixPost(x = x, n = n, par = parE,         weights = weights)    controlBetamixPost <- h_getBetamixPost(x = xS, n = nS, par = parS,         weights = weightsS)    assert_names(names(activeBetamixPost), identical.to = c("par",         "weights"))    assert_names(names(controlBetamixPost), identical.to = c("par",         "weights"))    if (relativeDelta) {        epsilon <- .Machine$double.xmin        integrand <- h_integrand_relDelta    }    else {        epsilon <- .Machine$double.xmin        integrand <- h_integrand    }    bounds <- h_get_bounds(controlBetamixPost = controlBetamixPost)    intRes <- integrate(f = integrand, lower = max(bounds[1],         ifelse(relativeDelta, 0, 0 - delta)), upper = min(ifelse(relativeDelta,         1, 1 - delta), bounds[2]), delta = delta, activeBetamixPost = activeBetamixPost,         controlBetamixPost = controlBetamixPost)    if (intRes$message == "OK") {        intRes$value    }    else {        stop(intRes$message)    }})(x = dots[[1L]][[1L]], n = 3L, xS = 4L, nS = 7L, delta = 0.15,     relativeDelta = FALSE, parE = c(a = 1, b = 1), parS = c(a = 1,     b = 1)): Assertion on 'xS' failed: Element 1 is not <= 3.
res2$oc
#> Error in eval(expr, envir, enclos): object 'res2' not found
# compared to res3, we see that there is no early futility stopping anymore,
# and the overall probability for stopping for futility (which is the type II error here)
# is much lower.

# Also, when deltaF is set to NULL, the futility decision is made when the probability to
# be deltaE better than the control is lower than 1 - tL. i.e. Pr(pE-pS>deltaE)<1-tL
res3 <- ocRctPostprobDist(
  nn = c(10, 20, 30), pE = 0.4, pS = 0.3, deltaE = 0.15, deltaF = NULL,
  relativeDelta = FALSE,
  tL = 0.2, tU = 0.8,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  randRatio = 1,
  ns = 100, nr = FALSE, d = NULL
)
#> Error in (function (x, n, xS = 0, nS = 0, delta, relativeDelta = FALSE,     parE = c(1, 1), weights, parS = c(1, 1), weightsS) {    if (is.vector(parE)) {        assert_true(identical(length(parE), 2L))        parE <- t(parE)    }    if (is.vector(parS)) {        assert_true(identical(length(parS), 2L))        parS <- t(parS)    }    if (missing(weights)) {        weights <- rep(1, nrow(parE))    }    if (missing(weightsS)) {        weightsS <- rep(1, nrow(parS))    }    assert_numeric(x, lower = 0, upper = n, finite = TRUE)    assert_number(n, lower = x, finite = TRUE)    assert_number(xS, lower = 0, upper = n, finite = TRUE)    assert_numeric(nS, lower = 0, finite = TRUE)    assert_number(delta, lower = 0, finite = TRUE)    assert_flag(relativeDelta)    assert_numeric(weights, lower = 0, finite = TRUE)    assert_numeric(weightsS, lower = 0, finite = TRUE)    assert_numeric(parE, lower = 0, finite = TRUE)    assert_numeric(parS, lower = 0, finite = TRUE)    activeBetamixPost <- h_getBetamixPost(x = x, n = n, par = parE,         weights = weights)    controlBetamixPost <- h_getBetamixPost(x = xS, n = nS, par = parS,         weights = weightsS)    assert_names(names(activeBetamixPost), identical.to = c("par",         "weights"))    assert_names(names(controlBetamixPost), identical.to = c("par",         "weights"))    if (relativeDelta) {        epsilon <- .Machine$double.xmin        integrand <- h_integrand_relDelta    }    else {        epsilon <- .Machine$double.xmin        integrand <- h_integrand    }    bounds <- h_get_bounds(controlBetamixPost = controlBetamixPost)    intRes <- integrate(f = integrand, lower = max(bounds[1],         ifelse(relativeDelta, 0, 0 - delta)), upper = min(ifelse(relativeDelta,         1, 1 - delta), bounds[2]), delta = delta, activeBetamixPost = activeBetamixPost,         controlBetamixPost = controlBetamixPost)    if (intRes$message == "OK") {        intRes$value    }    else {        stop(intRes$message)    }})(x = dots[[1L]][[1L]], n = 3L, xS = 4L, nS = 7L, delta = 0.15,     relativeDelta = FALSE, parE = c(a = 1, b = 1), parS = c(a = 1,     b = 1)): Assertion on 'xS' failed: Element 1 is not <= 3.
res3$oc
#> Error in eval(expr, envir, enclos): object 'res3' not found

# As when deltaF is set to null, it sometimes makes sense for a futility decision
# to be made even when the treatment is better than the control, e.g. when that difference
# is small. This can be accomplished by specifying a negative value for deltaF.
res4 <- ocRctPostprobDist(
  nn = c(10, 20, 30), pE = 0.4, pS = 0.3,
  deltaE = 0.15, deltaF = -0.05,
  relativeDelta = FALSE,
  tL = 0.2, tU = 0.8,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  randRatio = 1,
  ns = 100, nr = FALSE, d = NULL
)
#> Error in (function (x, n, xS = 0, nS = 0, delta, relativeDelta = FALSE,     parE = c(1, 1), weights, parS = c(1, 1), weightsS) {    if (is.vector(parE)) {        assert_true(identical(length(parE), 2L))        parE <- t(parE)    }    if (is.vector(parS)) {        assert_true(identical(length(parS), 2L))        parS <- t(parS)    }    if (missing(weights)) {        weights <- rep(1, nrow(parE))    }    if (missing(weightsS)) {        weightsS <- rep(1, nrow(parS))    }    assert_numeric(x, lower = 0, upper = n, finite = TRUE)    assert_number(n, lower = x, finite = TRUE)    assert_number(xS, lower = 0, upper = n, finite = TRUE)    assert_numeric(nS, lower = 0, finite = TRUE)    assert_number(delta, lower = 0, finite = TRUE)    assert_flag(relativeDelta)    assert_numeric(weights, lower = 0, finite = TRUE)    assert_numeric(weightsS, lower = 0, finite = TRUE)    assert_numeric(parE, lower = 0, finite = TRUE)    assert_numeric(parS, lower = 0, finite = TRUE)    activeBetamixPost <- h_getBetamixPost(x = x, n = n, par = parE,         weights = weights)    controlBetamixPost <- h_getBetamixPost(x = xS, n = nS, par = parS,         weights = weightsS)    assert_names(names(activeBetamixPost), identical.to = c("par",         "weights"))    assert_names(names(controlBetamixPost), identical.to = c("par",         "weights"))    if (relativeDelta) {        epsilon <- .Machine$double.xmin        integrand <- h_integrand_relDelta    }    else {        epsilon <- .Machine$double.xmin        integrand <- h_integrand    }    bounds <- h_get_bounds(controlBetamixPost = controlBetamixPost)    intRes <- integrate(f = integrand, lower = max(bounds[1],         ifelse(relativeDelta, 0, 0 - delta)), upper = min(ifelse(relativeDelta,         1, 1 - delta), bounds[2]), delta = delta, activeBetamixPost = activeBetamixPost,         controlBetamixPost = controlBetamixPost)    if (intRes$message == "OK") {        intRes$value    }    else {        stop(intRes$message)    }})(x = dots[[1L]][[1L]], n = 5L, xS = 0L, nS = 5L, delta = -0.05,     relativeDelta = FALSE, parE = c(a = 1, b = 1), parS = c(a = 1,     b = 1)): Assertion on 'delta' failed: Element 1 is not >= 0.
res4$oc
#> Error in eval(expr, envir, enclos): object 'res4' not found