Calculate operating characteristics for RCT against SOC, using the predictive probability method with beta priors

The randomization works as follows. According to the randomization ratio and the maximum sample size, patients are allocated to the treatment and SOC arms. The number of patients in the active treatment arm is rounded to the next higher integer. That is, the sequence of patients is determined from the start, such that the number of patients in both arms is constant across trial simulations.

Usage

ocRctPredprobDist(
  nn,
  pE,
  pS,
  delta = 0,
  deltaFu = delta,
  relativeDelta = FALSE,
  tT,
  tFu = 1 - tT,
  phiL = 1 - phiFu,
  phiU,
  phiFu = 1 - phiL,
  parE = c(a = 1, b = 1),
  parS = c(a = 1, b = 1),
  randRatio = 1,
  ns = 10000,
  nr = FALSE,
  d = NULL,
  nnF = nn
)

Arguments

nn

vector of look locations for efficacy . Note that the maximum sample size is derived as the maximum value (latest look) in this vector. (if futility looks should be different, please specify also nnF)

pE

true rate in the treatment population

pS

true rate in the standard of care population

delta

We want to be better than the control + delta (default: 0)

deltaFu

declare futility when worse than the control + deltaW. Specify only when decision rule 2 is used.

relativeDelta

see postprobDist

tT

threshold for the probability to be above control + delta at the end of the trial

tFu

threshold for the probability to be under control + deltaFu at the end of the trial. Specify only when decision rule 2 is used.

phiL

lower threshold on the predictive probability

phiU

upper threshold on the predictive probability

phiFu

threshold on the predictive probability for futility. Specify only when decision rule 2 is used. phiL argument should be skipped in this case.

parE

beta parameters for the prior on the control proportion

parS

beta parameters for the prior on the treatment proportion

randRatio

the randomization ratio (active vs. control) to be used. default: 1.

ns

number of simulations

nr

generate random look locations? (not default)

d

distance for random looks around the look locations in nn

nnF

vector of look locations for futility (default: same as efficacy)

Value

A list with the following elements: oc: matrix with operating characteristics (see Details section) Decision: vector of the decisions made in the simulated trials (TRUE for success, FALSE for failure) SampleSize: vector of the sample sizes in the simulated trials SampleSizeActive: vector of the patients with treatment in the simulated trials SampleSizeControl: vector of the patients with control in the simulated trials nn: vector of look locations nnE: vector of efficacy look locations nnF: vector of futility look locations todo: would we like to return nnr instead, the actual look locations? params: input parameters of this function

Details

Decision rule 1:The trial is stopped for efficacy if the predictive probability of a successful trial is larger than phiU, and stopped for futility if it is below phiL. A trial is successful if after the maximum number of patients the posterior probability of the treatment being at least delta better than the control is above tT. Otherwise the decision is "failure". Hence there is no gray zone.

Decision rule 2:A variation can be requested when skipping the argument phiL and utilizing the arguments p1, tFu & PhiFu. The trial can be stopped for futility if the predictive probability of an unsuccessful trial is larger than phiFu. In this case, the decision is "failure" when the posterior probability of having the treatment response rate at most p1 is above tFu at the maximum number of patients.

Returned operating characteristics in a matrix include: ExpectedN: expected number of patients in the trials ExpectedNactive: expected number of patients with treatment ExpectedNcontrol: expected number of patients with SOC PrStopEarly: probability to stop the trial early (before reaching the maximum sample size) PrEfficacy: probability to decide for efficacy PrFutility: probability to decide for futility PrGrayZone: probability of no decision at the end ("gray zone")

Examples

# similar input parameters as the oc.predprobDist function
# default simulation number is 10000. for illustration, the ns is set as 100 to
# save program running time. Please note that 10 simulation run (ns=10) may not
# be sufficient for a consistent OC evaluation.
# 1)
set.seed(20)
res <- ocRctPredprobDist(
  nn = c(10, 20, 30), pE = 0.4, pS = 0.3, delta = 0.1,
  relativeDelta = FALSE, tT = 0.6,
  phiL = 0.2, phiU = 0.8,
  parE = c(a = 1, b = 1), parS = c(a = 1, b = 1),
  randRatio = 1,
  ns = 10, nr = FALSE, d = NULL
)

res$oc
#>      ExpectedN ExpectedNactive ExpectedNcontrol PrStopEarly PrEarlyEff
#> [1,]        19               9               10         0.6        0.1
#>      PrEarlyFut PrEfficacy PrFutility PrGrayZone
#> [1,]        0.5        0.3        0.7          0

# separate specification of efficacy and futility analyses can be performed. E.g.
# here have futility decisions only at the final analysis:
set.seed(20)
res1 <- ocRctPredprobDist(
  nn = c(10, 20, 30), pE = 0.4, pS = 0.3, delta = 0.1, relativeDelta = FALSE,
  tT = 0.6,
  phiL = 0.2, phiU = 0.8,
  parE = c(a = 1, b = 1), parS = c(a = 1, b = 1),
  randRatio = 1,
  ns = 10, nr = FALSE, d = NULL, nnF = c(30)
)

res1$oc
#>      ExpectedN ExpectedNactive ExpectedNcontrol PrStopEarly PrEarlyEff
#> [1,]        28            13.9             14.1         0.1        0.1
#>      PrEarlyFut PrEfficacy PrFutility PrGrayZone
#> [1,]          0        0.4        0.6          0

# 2)
# to examine operating characters of a phase Ib design with a
# seperate rule for futility
# at final analysis
# denote treatment response rate as RRS
# trial success is defined as: P(RRS>control+delta)>tT
# trial failure is defined as: P(RRS<control+deltaFu)>tFu
# (gray zone could occur in the final analysis)

set.seed(20)
res2 <- ocRctPredprobDist(
  nn = c(10, 20, 30), pE = 0.4, pS = 0.3, delta = 0.1, deltaFu = 0.05,
  relativeDelta = FALSE,
  tT = 0.6, tFu = 0.6,
  phiFu = 0.8, phiU = 0.8,
  parE = c(a = 1, b = 1), parS = c(a = 1, b = 1),
  randRatio = 1,
  ns = 10, nr = FALSE, d = NULL
)

res2$oc
#>      ExpectedN ExpectedNactive ExpectedNcontrol PrStopEarly PrEarlyEff
#> [1,]        24            11.8             12.2         0.4        0.1
#>      PrEarlyFut PrEfficacy PrFutility PrGrayZone
#> [1,]        0.3        0.4        0.4        0.2