Compute the predictive probability that the trial will be successful, with a fixed response rate threshold

Compute the predictive probability of trial success given current data. Success means that at the end of the trial the posterior probability Pr(P_E > p) >= thetaT, where p is the fixed response rate threshold. Then the predictive probability for success is: pp = sum over i: Pr(Y=i|x,n)*IPr(P_E > p|x,Y=i)>=thetaT where Y is the number of future responses in the treatment group and x is the current number of responses in the treatment group (out of n). The prior is P_E ~ beta(a, b), default uniform which is a beta(1,1). However, also a beta mixture prior can be specified.

Usage

predprob(x, n, Nmax, p, thetaT, parE = c(1, 1), weights)

Arguments

x

number of successes

n

number of patients

Nmax

maximum number of patients at the end of the trial

p

threshold on the response rate

thetaT

threshold on the probability to be above p

parE

the beta parameters matrix, with K rows and 2 columns, corresponding to the beta parameters of the K components. default is a uniform prior.

weights

the mixture weights of the beta mixture prior. Default are uniform weights across mixture components.

Value

The predictive probability, a numeric value. In addition, a list called table is returned as attribute of the returned number.

Details

A table with the following contents will be included in the table attribute of the return value: i: Y=i (number of future successes in Nmax-n subjects) py: Pr(Y=i|x) using beta-(mixture)-binomial distribution b: Pr(P_E > p | x, Y=i) using beta (mixture) posterior bgttheta: indicator I(b>thetaT)

References

Lee, J. J., & Liu, D. D. (2008). A predictive probability design for phase II cancer clinical trials. Clinical Trials, 5(2), 93-106. doi:10.1177/1740774508089279

Examples

## The original Lee and Liu (Table 1) example:
## Nmax=40, x=16, n=23, beta(0.6,0.4) prior distribution,
## thetaT=0.9. The control response rate is 60%:
predprob(
  x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.9,
  parE = c(0.6, 0.4)
)
#> [1] 0.5655589
#> attr(,"table")
#>        i     py      b bgttheta
#>  [1,]  0 0.0000 0.0059        0
#>  [2,]  1 0.0000 0.0138        0
#>  [3,]  2 0.0001 0.0296        0
#>  [4,]  3 0.0006 0.0581        0
#>  [5,]  4 0.0021 0.1049        0
#>  [6,]  5 0.0058 0.1743        0
#>  [7,]  6 0.0135 0.2679        0
#>  [8,]  7 0.0276 0.3821        0
#>  [9,]  8 0.0497 0.5085        0
#> [10,]  9 0.0794 0.6349        0
#> [11,] 10 0.1129 0.7489        0
#> [12,] 11 0.1426 0.8415        0
#> [13,] 12 0.1587 0.9089        1
#> [14,] 13 0.1532 0.9528        1
#> [15,] 14 0.1246 0.9781        1
#> [16,] 15 0.0811 0.9910        1
#> [17,] 16 0.0381 0.9968        1
#> [18,] 17 0.0099 0.9990        1
## So the predictive probability is 56.6%. 12 responses
## in the remaining 17 patients are required for success.

## Lowering/Increasing the probability threshold thetaT of course increases
## /decreases the predictive probability of success, respectively:
predprob(
  x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.8,
  parE = c(0.6, 0.4)
)
#> [1] 0.7081907
#> attr(,"table")
#>        i     py      b bgttheta
#>  [1,]  0 0.0000 0.0059        0
#>  [2,]  1 0.0000 0.0138        0
#>  [3,]  2 0.0001 0.0296        0
#>  [4,]  3 0.0006 0.0581        0
#>  [5,]  4 0.0021 0.1049        0
#>  [6,]  5 0.0058 0.1743        0
#>  [7,]  6 0.0135 0.2679        0
#>  [8,]  7 0.0276 0.3821        0
#>  [9,]  8 0.0497 0.5085        0
#> [10,]  9 0.0794 0.6349        0
#> [11,] 10 0.1129 0.7489        0
#> [12,] 11 0.1426 0.8415        1
#> [13,] 12 0.1587 0.9089        1
#> [14,] 13 0.1532 0.9528        1
#> [15,] 14 0.1246 0.9781        1
#> [16,] 15 0.0811 0.9910        1
#> [17,] 16 0.0381 0.9968        1
#> [18,] 17 0.0099 0.9990        1
## 70.8%
predprob(
  x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.95,
  parE = c(0.6, 0.4)
)
#> [1] 0.4068235
#> attr(,"table")
#>        i     py      b bgttheta
#>  [1,]  0 0.0000 0.0059        0
#>  [2,]  1 0.0000 0.0138        0
#>  [3,]  2 0.0001 0.0296        0
#>  [4,]  3 0.0006 0.0581        0
#>  [5,]  4 0.0021 0.1049        0
#>  [6,]  5 0.0058 0.1743        0
#>  [7,]  6 0.0135 0.2679        0
#>  [8,]  7 0.0276 0.3821        0
#>  [9,]  8 0.0497 0.5085        0
#> [10,]  9 0.0794 0.6349        0
#> [11,] 10 0.1129 0.7489        0
#> [12,] 11 0.1426 0.8415        0
#> [13,] 12 0.1587 0.9089        0
#> [14,] 13 0.1532 0.9528        1
#> [15,] 14 0.1246 0.9781        1
#> [16,] 15 0.0811 0.9910        1
#> [17,] 16 0.0381 0.9968        1
#> [18,] 17 0.0099 0.9990        1
## 40.7%

## Instead of a fixed beta prior on the response rate, dynamic borrowing
## from a previous data set is possible by using a beta-mixture prior.
## For example, assume we have a previous trial where we saw 25 responses
## in 40 patients. We would like to use information worth 10 patients in our
## trial, but be robust against deviations from that response rate (maybe
## because it was conducted in slightly different disease or patients).
## Then we can use a beta mixture prior, with the informative component getting
## weight of 1/4:
predprob(
  x = 20, n = 23, Nmax = 40, p = 0.6, thetaT = 0.9,
  parE = rbind(c(1, 1), c(25, 15)),
  weights = c(3, 1)
)
#> [1] 0.9874431
#> attr(,"table")
#>        i     py      b bgttheta
#>  [1,]  0 0.0000 0.1560        0
#>  [2,]  1 0.0000 0.2352        0
#>  [3,]  2 0.0000 0.3303        0
#>  [4,]  3 0.0000 0.4359        0
#>  [5,]  4 0.0001 0.5449        0
#>  [6,]  5 0.0003 0.6500        0
#>  [7,]  6 0.0011 0.7447        0
#>  [8,]  7 0.0031 0.8244        0
#>  [9,]  8 0.0080 0.8871        0
#> [10,]  9 0.0177 0.9327        1
#> [11,] 10 0.0348 0.9634        1
#> [12,] 11 0.0613 0.9821        1
#> [13,] 12 0.0973 0.9923        1
#> [14,] 13 0.1394 0.9972        1
#> [15,] 14 0.1783 0.9991        1
#> [16,] 15 0.1964 0.9998        1
#> [17,] 16 0.1708 1.0000        1
#> [18,] 17 0.0913 1.0000        1
## Since the response rate in the historical dataset is lower (62.5% < 69.6%)
## than in the trial, but similar, we notice that the predictive probability
## of success is now lower.