Compute the predictive probability that the trial will be successful, with a prior distribution on the SOC

Compute the predictive probability of trial success given current data. Success means that at the end of the trial the posterior probability Pr(P_E > P_S + delta_0 | data) >= thetaT. Then the predictive probability for success is: pp = sum over i: Pr(Y = i | x , n)*I{Pr(P_E > P_S + delta | x,Y=i) >= thetaT}, where Y is the number of future responses in the treatment group and x is the current number of responses in the treatment group out of n. Prior is P_E ~ beta(a, b) and uniform which is a beta(1,1). However, also a beta mixture prior can be specified. Analogously for P_S either a classic beta prior or a beta mixture prior can be specified.

Also data on the SOC might be available. Then the predictive probability is more generally defined as pp = sum over i, j: Pr(Y = i | x, n)*Pr(Z = j | xS, nS )*I{Pr(P_E > P_S + delta | x,xS, Y=i, Z=j ) >= thetaT} where Z is the future number of responses in the SOC group, and xS is the current number of responses in the SOC group.

In the case where NmaxControl = 0, a table with the following contents will be included in the return output :

i: Y = i, number of future successes in Nmax-n subjects.

density: Pr(Y = i|x) using beta-(mixture)-binomial distribution.

posterior: Pr(P_E > P_S + delta | x, Y = i) using beta posterior.

success: indicator I(b>thetaT).

A table with the following contents will be included in the return value (in the case that NmaxControl is zero): i: Y=i (number of future successes in Nmax-n subjects) py: Pr(Y=i|x) using beta-(mixture)-binomial distribution b: Pr(P_E > P_S + delta | x, Y=i) bgttheta: indicator I(b>thetaT)

If NmaxControl is not zero, i.e., when data on the control treatment is available in this trial, then a list with will be included with the following elements: pyz: matrix with the probabilities Pr(Y=i, Z=j | x, xS) b: matrix with Pr(P_E > P_S + delta | x, xS, Y=i, Z=j) bgttheta: matrix of indicators I(b>thetaT)

Usage

predprobDist(
  x,
  n,
  xS = 0,
  nS = 0,
  Nmax,
  NmaxControl = 0,
  delta = 0,
  relativeDelta = FALSE,
  parE = c(a = 1, b = 1),
  weights,
  parS = c(a = 1, b = 1),
  weightsS,
  thetaT
)

Arguments

x

(number):
number of successes in the treatment group at interim

n

(number):
number of patients in the treatment group at interim

xS

(number):
number of successes in the SOC group at interim

nS

(number):
number of patients in the SOC group

Nmax

(number):
maximum number of patients in the treatment group at the end of the trial

NmaxControl

(number):
maximum number of patients at the end of the trial in the SOC group

delta

(``):
margin by which the response rate in the treatment group should be better than in the SOC group

relativeDelta

(``):
see [postprobDist()]

parE

(``):
the beta parameters matrix, with K rows and 2 columns, corresponding to the beta parameters of the K components. default is a uniform prior.

weights

(``):
the mixture weights of the beta mixture prior. Default are uniform weights across mixture components.

parS

(``):
beta prior parameters in the SOC group

weightsS

(``):
weights for the SOC group

thetaT

(``):
threshold on the probability to be used

Value

A list is returned with names result for predictive probability and table (single arm case) or list tables with ... (to be completed)

Delta from postprobDist :

The desired improvement is denoted as delta. There are two options in using delta. The absolute case when relativeDelta = FALSE and relative as when relativeDelta = TRUE.

The absolute case is when we define an absolute delta, greater than P_S, the response rate of the SOC group such that the posterior is Pr(P_E > P_S + delta | data).

In the relative case, we suppose that the treatment group's response rate is assumed to be greater than P_S + (1-P_S) * delta such that the posterior is Pr(P_E > P_S + (1 - P_S) * delta | data).

References

Lee, J. J., & Liu, D. D. (2008). A predictive probability design for phase II cancer clinical trials. Clinical Trials, 5(2), 93-106. doi:10.1177/1740774508089279

Examples

# The original Lee and Liu (Table 1) example:
# Nmax=40, x=16, n=23, beta(0.6,0.4) prior distribution,
# thetaT=0.9. The control response rate is 60%:
predprob(
  x = 16, n = 23, Nmax = 40, p = 0.6, thetaT = 0.9,
  parE = c(0.6, 0.4)
)
#> $result
#> [1] 0.5655589
#> 
#> $table
#>    counts cumul_counts density   posterior success
#> 1       0           23  0.0000 0.005858397   FALSE
#> 2       1           24  0.0000 0.013782046   FALSE
#> 3       2           25  0.0001 0.029584325   FALSE
#> 4       3           26  0.0006 0.058130376   FALSE
#> 5       4           27  0.0021 0.104881818   FALSE
#> 6       5           28  0.0058 0.174328134   FALSE
#> 7       6           29  0.0135 0.267887754   FALSE
#> 8       7           30  0.0276 0.382146405   FALSE
#> 9       8           31  0.0497 0.508508727   FALSE
#> 10      9           32  0.0794 0.634871049   FALSE
#> 11     10           33  0.1129 0.748893301   FALSE
#> 12     11           34  0.1426 0.841482798   FALSE
#> 13     12           35  0.1587 0.908912106    TRUE
#> 14     13           36  0.1532 0.952764733    TRUE
#> 15     14           37  0.1246 0.978098514    TRUE
#> 16     15           38  0.0811 0.991013775    TRUE
#> 17     16           39  0.0381 0.996776597    TRUE
#> 18     17           40  0.0099 0.999003945    TRUE
#> 
# So the predictive probability is 56.6%. 12 responses
# in the remaining 17 patients are required for success.

# Now to the modified example, where the control response
# rate p has a beta(7, 11) distribution.
# For example if the 60% threshold (control) stems from a trial
# with 10 patients, of which 6 were observed with response.
predprobDist(
  x = 16, n = 23, Nmax = 40, delta = 0, thetaT = 0.9,
  parE = c(0.6, 0.4), parS = c(7, 11)
)
#> $result
#> [1] 0.9778204
#> 
#> $table
#>    counts      density posterior success
#> 1       0 1.885196e-06 0.5530716   FALSE
#> 2       1 2.273514e-05 0.6214508   FALSE
#> 3       2 1.429066e-04 0.6856889   FALSE
#> 4       3 6.210428e-04 0.7444267   FALSE
#> 5       4 2.088409e-03 0.7967127   FALSE
#> 6       5 5.765730e-03 0.8420189   FALSE
#> 7       6 1.353693e-02 0.8802192   FALSE
#> 8       7 2.762957e-02 0.9115393    TRUE
#> 9       8 4.969953e-02 0.9364854    TRUE
#> 10      9 7.939015e-02 0.9557619    TRUE
#> 11     10 1.129104e-01 0.9701884    TRUE
#> 12     11 1.426318e-01 0.9806229    TRUE
#> 13     12 1.587354e-01 0.9878972    TRUE
#> 14     13 1.531657e-01 0.9927688    TRUE
#> 15     14 1.245523e-01 0.9958900    TRUE
#> 16     15 8.109152e-02 0.9977932    TRUE
#> 17     16 3.813232e-02 0.9988905    TRUE
#> 18     17 9.881666e-03 0.9994836    TRUE
#> 
# Now only7 responses are needed in the remaining 17 patients
# in order to have a successful trial.
# Also, the predictive probability for a Go decision
# is now 97.78% instead of merely 56.6%.

# If the response threshold of 60% stems from an SOC estimate (e.g. 50% = 5/10)
# and a margin of e.g. 10%, then the margin can be specified with
# the delta argument:
predprobDist(
  x = 16, n = 23, Nmax = 40, delta = 0.1, thetaT = 0.9,
  parE = c(1, 1), parS = c(6, 11)
)
#> $result
#> [1] 0.8004621
#> 
#> $table
#>    counts      density posterior success
#> 1       0 2.283242e-06 0.3716084   FALSE
#> 2       1 2.749403e-05 0.4400664   FALSE
#> 3       2 1.721366e-04 0.5092459   FALSE
#> 4       3 7.433170e-04 0.5772595   FALSE
#> 5       4 2.477723e-03 0.6423841   FALSE
#> 6       5 6.764185e-03 0.7031617   FALSE
#> 7       6 1.566443e-02 0.7584690   FALSE
#> 8       7 3.145318e-02 0.8075520   FALSE
#> 9       8 5.550560e-02 0.8500272   FALSE
#> 10      9 8.672751e-02 0.8858553   FALSE
#> 11     10 1.202621e-01 0.9152925    TRUE
#> 12     11 1.475944e-01 0.9388274    TRUE
#> 13     12 1.589479e-01 0.9571117    TRUE
#> 14     13 1.477400e-01 0.9708907    TRUE
#> 15     14 1.151221e-01 0.9809398    TRUE
#> 16     15 7.137569e-02 0.9880117    TRUE
#> 17     16 3.172253e-02 0.9927963    TRUE
#> 18     17 7.697378e-03 0.9958932    TRUE
#> 
# This again changes the result: Only 10 future responses are required,
# the predictive probability of success is now 80%.

# Next we will use a beta mixture prior, with a weight of 1/3 for an
# informative beta(50, 10) prior, on the experimental response rate:
predprobDist(
  x = 16, n = 23, Nmax = 40, delta = 0.1, thetaT = 0.9,
  parE = rbind(c(1, 1), c(50, 10)),
  weights = c(2, 1), parS = c(6, 11)
)
#> $result
#> [1] 0.8719377
#> 
#> $table
#>    counts      density posterior success
#> 1       0 0.0000013665 0.3716442   FALSE
#> 2       1 0.0000164568 0.4401569   FALSE
#> 3       2 0.0001030642 0.5094585   FALSE
#> 4       3 0.0004453712 0.5777253   FALSE
#> 5       4 0.0014870715 0.6433349   FALSE
#> 6       5 0.0040748393 0.7049694   FALSE
#> 7       6 0.0095102819 0.7616623   FALSE
#> 8       7 0.0193925232 0.8127701   FALSE
#> 9       8 0.0352153786 0.8578603   FALSE
#> 10      9 0.0578159915 0.8965520   FALSE
#> 11     10 0.0867552786 0.9284286    TRUE
#> 12     11 0.1194156177 0.9531859    TRUE
#> 13     12 0.1495254621 0.9710045    TRUE
#> 14     13 0.1661334169 0.9828056    TRUE
#> 15     14 0.1566053804 0.9900552    TRUE
#> 16     15 0.1165762124 0.9942733    TRUE
#> 17     16 0.0605313683 0.9966617    TRUE
#> 18     17 0.0163949189 0.9980136    TRUE
#> 
# Now still 10 future responses are required for success, but the predictive
# success probability is higher (87.2%).

# We can also have additional controls in our trial. For example
# assume that in the interim analysis we had 10 control patients observed
# with 5 responses out of 20 in total at the end.
# We also had historical control data (say 60 patients of
# which 20 responded), but in order to be robust against changes in the response
# rate over time, we will again use a beta mixture, putting only 1/3 of weight
# on the historical controls:
predprobDist(
  x = 16, n = 23, xS = 5, nS = 10,
  Nmax = 40, NmaxControl = 20,
  delta = 0.1, thetaT = 0.9,
  parE = rbind(c(1, 1), c(50, 10)),
  weights = c(2, 1),
  parS = rbind(c(1, 1), c(20, 40)),
  weightsS = c(2, 1)
)
#> $result
#> [1] 0.5989817
#> 
#> $tables
#> $tables$pyz
#>               0            1            2            3            4
#> 0  1.636390e-08 7.156485e-08 1.591139e-07 2.356831e-07 2.610416e-07
#> 1  1.970710e-07 8.618577e-07 1.916213e-06 2.838338e-06 3.143732e-06
#> 2  1.234198e-06 5.397564e-06 1.200069e-05 1.777569e-05 1.968828e-05
#> 3  5.333342e-06 2.332450e-05 5.185858e-05 7.681409e-05 8.507897e-05
#> 4  1.780775e-05 7.787929e-05 1.731531e-04 2.564783e-04 2.840743e-04
#> 5  4.879639e-05 2.134030e-04 4.744702e-04 7.027959e-04 7.784138e-04
#> 6  1.138861e-04 4.980621e-04 1.107368e-03 1.640258e-03 1.816743e-03
#> 7  2.322264e-04 1.015604e-03 2.258046e-03 3.344668e-03 3.704541e-03
#> 8  4.217058e-04 1.844261e-03 4.100443e-03 6.073669e-03 6.727170e-03
#> 9  6.923492e-04 3.027876e-03 6.732036e-03 9.971643e-03 1.104455e-02
#> 10 1.038899e-03 4.543453e-03 1.010170e-02 1.496286e-02 1.657280e-02
#> 11 1.430008e-03 6.253904e-03 1.390463e-02 2.059586e-02 2.281189e-02
#> 12 1.790574e-03 7.830784e-03 1.741059e-02 2.578896e-02 2.856375e-02
#> 13 1.989455e-03 8.700558e-03 1.934441e-02 2.865337e-02 3.173636e-02
#> 14 1.875357e-03 8.201566e-03 1.823497e-02 2.701005e-02 2.991622e-02
#> 15 1.396006e-03 6.105202e-03 1.357402e-02 2.010614e-02 2.226948e-02
#> 16 7.248660e-04 3.170083e-03 7.048212e-03 1.043997e-02 1.156327e-02
#> 17 1.963299e-04 8.586168e-04 1.909008e-03 2.827665e-03 3.131910e-03
#>               5            6            7            8            9
#> 0  2.326759e-07 1.759661e-07 1.156659e-07 6.436294e-08 2.734766e-08
#> 1  2.802123e-06 2.119166e-06 1.392968e-06 7.751249e-07 3.293487e-07
#> 2  1.754888e-05 1.327172e-05 8.723754e-06 4.854382e-06 2.062615e-06
#> 3  7.583400e-05 5.735108e-05 3.769797e-05 2.097724e-05 8.913179e-06
#> 4  2.532058e-04 1.914923e-04 1.258716e-04 7.004193e-05 2.976064e-05
#> 5  6.938287e-04 5.247228e-04 3.449104e-04 1.919273e-04 8.154943e-05
#> 6  1.619329e-03 1.224652e-03 8.049875e-04 4.479398e-04 1.903285e-04
#> 7  3.301993e-03 2.497203e-03 1.641459e-03 9.133991e-04 3.881010e-04
#> 8  5.996173e-03 4.534734e-03 2.980768e-03 1.658665e-03 7.047626e-04
#> 9  9.844412e-03 7.445047e-03 4.893772e-03 2.723167e-03 1.157067e-03
#> 10 1.477195e-02 1.117160e-02 7.343307e-03 4.086224e-03 1.736226e-03
#> 11 2.033307e-02 1.537732e-02 1.010781e-02 5.624545e-03 2.389855e-03
#> 12 2.545992e-02 1.925461e-02 1.265642e-02 7.042736e-03 2.992441e-03
#> 13 2.828778e-02 2.139324e-02 1.406218e-02 7.824980e-03 3.324815e-03
#> 14 2.666542e-02 2.016630e-02 1.325569e-02 7.376204e-03 3.134131e-03
#> 15 1.984960e-02 1.501168e-02 9.867467e-03 5.490807e-03 2.333030e-03
#> 16 1.030676e-02 7.794710e-03 5.123612e-03 2.851062e-03 1.211409e-03
#> 17 2.791586e-03 2.111197e-03 1.387730e-03 7.722102e-04 3.281102e-04
#>              10
#> 0  6.714161e-09
#> 1  8.085884e-08
#> 2  5.063954e-07
#> 3  2.188287e-06
#> 4  7.306577e-06
#> 5  2.002131e-05
#> 6  4.672782e-05
#> 7  9.528322e-05
#> 8  1.730272e-04
#> 9  2.840731e-04
#> 10 4.262634e-04
#> 11 5.867366e-04
#> 12 7.346783e-04
#> 13 8.162799e-04
#> 14 7.694648e-04
#> 15 5.727855e-04
#> 16 2.974148e-04
#> 17 8.055479e-05
#> 
#> $tables$b
#>            0         1         2         3         4         5          6
#> 0  0.5257346 0.4312132 0.3464980 0.2680767 0.1946083 0.1278211 0.07269905
#> 1  0.6164888 0.5248360 0.4366781 0.3493814 0.2622247 0.1781748 0.10497694
#> 2  0.7011371 0.6164236 0.5290576 0.4365535 0.3381350 0.2374450 0.14489916
#> 3  0.7759524 0.7012532 0.6186670 0.5251436 0.4190636 0.3038799 0.19209595
#> 4  0.8386592 0.7757668 0.7011830 0.6107733 0.5013600 0.3751874 0.24579597
#> 5  0.8885541 0.8379634 0.7735289 0.6898140 0.5816063 0.4489661 0.30507234
#> 6  0.9262891 0.8874080 0.8341266 0.7598142 0.6570959 0.5231075 0.36911214
#> 7  0.9534412 0.9249253 0.8827878 0.8195885 0.7260451 0.5960170 0.43736984
#> 8  0.9720282 0.9521299 0.9203481 0.8689964 0.7874849 0.6665275 0.50942884
#> 9  0.9841041 0.9709592 0.9482116 0.9085447 0.8409019 0.7334844 0.58443319
#> 10 0.9915057 0.9833341 0.9679771 0.9390206 0.8858537 0.7952108 0.66022457
#> 11 0.9957471 0.9909860 0.9812476 0.9613482 0.9218925 0.8493818 0.73288679
#> 12 0.9980000 0.9953960 0.9895849 0.9766815 0.9489523 0.8937678 0.79763325
#> 13 0.9991074 0.9977573 0.9944615 0.9864822 0.9678227 0.9274362 0.85085818
#> 14 0.9996155 0.9989408 0.9971338 0.9923469 0.9801227 0.9512658 0.89163037
#> 15 0.9998370 0.9995057 0.9985292 0.9956911 0.9877617 0.9673400 0.92141809
#> 16 0.9999308 0.9997681 0.9992393 0.9975518 0.9923973 0.9779514 0.94275092
#> 17 0.9999701 0.9998893 0.9995995 0.9985857 0.9952127 0.9849739 0.95808432
#>             7          8           9          10
#> 0  0.03446032 0.01337663 0.004262331 0.001121434
#> 1  0.05165626 0.02087630 0.006951808 0.001920433
#> 2  0.07407344 0.03122238 0.010897011 0.003173482
#> 3  0.10214142 0.04501075 0.016525195 0.005098255
#> 4  0.13617138 0.06292971 0.024413350 0.008024192
#> 5  0.17649408 0.08586691 0.035385174 0.012466593
#> 6  0.22363510 0.11505762 0.050652910 0.019241758
#> 7  0.27841864 0.15220229 0.071967427 0.029613131
#> 8  0.34180117 0.19937000 0.101642262 0.045393701
#> 9  0.41418768 0.25838321 0.142170641 0.068807293
#> 10 0.49421282 0.32949245 0.195144167 0.101833237
#> 11 0.57771556 0.40988163 0.259716644 0.145041894
#> 12 0.65829416 0.49352938 0.331871554 0.196700263
#> 13 0.72980920 0.57339337 0.405739215 0.253236958
#> 14 0.78891577 0.64439909 0.476191446 0.310932082
#> 15 0.83555624 0.70474947 0.540514258 0.367368500
#> 16 0.87168673 0.75516790 0.598295826 0.421730160
#> 17 0.89979569 0.79748034 0.650440254 0.474339349
#> 
#> $tables$success
#>        0     1     2     3     4     5     6     7     8     9    10
#> 0  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 1  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 2  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 3  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 4  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 5  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 6   TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 7   TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 8   TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 9   TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 10  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 11  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#> 12  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#> 13  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE
#> 14  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE
#> 15  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE
#> 16  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE
#> 17  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE
#> 
#> 
# Now we only have 59.9% predictive probability of success. The bgttheta matrix
# lists the cases where we would have success, for all possible combinations
# of future responses in the control and experimental arms. We see that the
# success threshold for experimental responses depends on the number of
# control responses.